// icedt-admin-angular/src/app/types/auth.types.ts
export interface LoginRequest {
  username: string;
  password: string;
}

export interface User {
  id: number;
  username: string;
  email: string;
  role: string;
}

export interface AuthResponse {
  isSuccess: boolean;
  message: string;
  token: string | null;
  refreshToken?: string | null;
  user?: User | null;
}